Add meta tags for search engine optimization, refactor sidebar to separate component
