+++ src/app/components/Experience.js
@@ -1,3 +1,4 @@
+'use client';
 import React, { useState, useEffect, useRef, useMemo } from 'react';
 import dynamic from 'next/dynamic';
 import 'react-quill-new/dist/quill.snow.css';
